Sparrow

Sparrow is an impementation of boosting that is optimized for training on very large datasets, as well as training in limited memory settings. The opimizations involve two technologies: early stopping and selective sampling. Please read our paper for more details.

0.1 Prerequisites

Install Rust following the instruction on the Rust offical webiste.

Compile Sparrow: cargo build --release

The Sparrow binary file would be generated at target/release/sparrow.

0.2 Usage

Sparrow is written as a Rust library. It also supports running as a binary. The sparrow binary reads configuration from a specified configuration file. Many examples of the configuration files can be found in the examples/ directory.

To run the Sparrow binary, please provide the path to the configuration file.

For training,

./sparrow train <path to the config file>

For testing (or prediction),

./sparrow test <path to the config file>

To use Sparrow as a Rust library, please read the document here.