Module sparrow::scanner::booster

The implementation of the AdaBoost algorithm with early stopping rule.

Modules

learner

learner_helpers

learner_stats

tree_node

Structs

Boosting
The boosting algorithm. It contains two functions, one for starting the network communication, the other for starting the training procedure.

Constants

MODEL_SIG_PLACEHOLDER

VERBOSE