Module sparrow::scanner::booster::learner

Structs

Learner
Statisitics of all weak rules that are being evaluated. The objective of Learner is to find a weak rule that satisfies the condition of the stopping rule.

TreeNode
A weak rule with an edge larger or equal to the targetting value of gamma

Constants

NUM_PREDS

PREDS

Type Definitions

RuleStats