Module sparrow::scanner::buffer_loader

A data loader with two independent caches. Alternatively, we use one of the caches to feed data to the boosting algorithm, and the other to load next sample set.

Structs

BufferLoader
Double-buffered sample set. It consists of two buffers stores in memory. One of the buffer is used for providing examples to the boosting algorithm, meanwhile the other would receive new examples from the sampler module. Once the other buffer is filled, the functions of the two buffers switch so that the sample set used for training would be updated periodically.