Struct sparrow::Config

pub struct Config {
    pub training_filename: String,
    pub num_examples: usize,
    pub num_features: usize,
    pub range: Range<usize>,
    pub positive: String,
    pub testing_filename: String,
    pub num_testing_examples: usize,
    pub max_sample_size: usize,
    pub max_bin_size: usize,
    pub min_gamma: f32,
    pub default_gamma: f32,
    pub max_trials_before_shrink: u32,
    pub min_ess: f32,
    pub num_iterations: usize,
    pub max_leaves: usize,
    pub channel_size: usize,
    pub buffer_size: usize,
    pub batch_size: usize,
    pub serial_sampling: bool,
    pub num_examples_per_block: usize,
    pub disk_buffer_filename: String,
    pub num_assigners: usize,
    pub num_samplers: usize,
    pub network: Vec<String>,
    pub port: u16,
    pub local_name: String,
    pub save_process: bool,
    pub save_interval: usize,
    pub debug_mode: bool,
    pub models_table_filename: String,
    pub incremental_testing: bool,
    pub testing_scores_only: bool,
}

Configuration for training and testing with Sparrow

Fields
training_filename: String

File path to the training data

num_examples: usize

Number of training examples

num_features: usize

Number of features

range: Range<usize>

Range of the features for creating weak rules

positive: String

Label for positive examples

testing_filename: String

File path to the testing data

num_testing_examples: usize

Number of testing examples

max_sample_size: usize

Number of examples to scan for generating heuristic used in Sparrow

max_bin_size: usize

Maximum number of bins for discretizing continous feature values

min_gamma: f32

Minimum value of the \gamma of the generated tree nodes

default_gamma: f32

Default maximum value of the \gamma for generating tree nodes

max_trials_before_shrink: u32

Maximum number of examples to scan before shrinking the value of \gamma

min_ess: f32

Minimum effective sample size for triggering resample

num_iterations: usize

Number of boosting iterations

max_leaves: usize

Maximum number of tree leaves in each boosted tree

channel_size: usize

Maximum number of elements in the channel connecting scanner and sampler

buffer_size: usize

Number of examples in the sample set that needs to be loaded into memory

batch_size: usize

Number of examples to process in each weak rule updates

serial_sampling: bool

Set to true to stop running sampler in the background of the scanner

num_examples_per_block: usize

Number of examples in a block on the stratified binary file

disk_buffer_filename: String

File name for the stratified binary file

num_assigners: usize

Number of threads for putting examples back to correct strata

num_samplers: usize

Number of threads for sampling examples from strata

network: Vec<String>

IP addresses of other machines in the network

port: u16

The network port used for parallel training

local_name: String

Identifier for the local machine

save_process: bool

Flag for keeping all intermediate models during training (for debugging purpose)

save_interval: usize

Number of iterations between persisting models on disk

debug_mode: bool

Flag for activating debug mode

models_table_filename: String

(for validation only) the file names of the models to run the validation

incremental_testing: bool

Flag indicating if models are trained incrementally

testing_scores_only: bool

Flag for validation mode, set to true to output raw scores of testing examples, and set to false for printing the validation scores but not raw scores